pyproject.toml
[project]
name = "backend-daily-jokes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "python-dotenv>=1.0.1",
]
